refactor(inspection-service): extract inspection-core sub-service; facade done
Move the residual CRUD + lifecycle (listInspections / getStats / computePreflight
/ getInspection / createInspection / createReinspection / getReinspectCandidates
/ applyServicePriceOverrides / createFromWizard / cloneInspection) plus the
recipient + people aggregation cards into InspectionCoreService. cloneInspection
calls getInspection/createInspection internally (same service). With this,
InspectionService is a pure delegation facade — every public method forwards to
a focused sub-service with a byte-identical moved body and an unchanged
signature; the constructor + named exports are unchanged so all callers, DI,
and the ~1937-test suite are untouched.
